refactor(auth): encapsulate email dialog in EmailLogin
Rename extracted email form component to EmailLogin and move dialog trigger/content into the same file.
