+++ src/app/_components/login-button.tsx
@@ -1,14 +1,11 @@
 'use client';
 
-import { useState } from 'react';
 import {
 	FaDiscord as DiscordIcon,
-	FaEnvelope as EmailIcon,
 	FaGoogle as GoogleIcon,
 } from 'react-icons/fa6';
 
 import { Button } from '~/components/ui/button';
-import { Dialog, DialogContent, DialogTrigger } from '~/components/ui/dialog';
 import {
 	DropdownMenu,
 	DropdownMenuContent,
@@ -17,11 +14,9 @@ import {
 } from '~/components/ui/dropdown-menu';
 import { authClient } from '~/lib/auth-client';
 
-import { EmailLoginForm } from './email-login-form';
+import { EmailLogin } from './email-login';
 
 export const LoginButton = () => {
-	const [open, setOpen] = useState(false);
-
 	const login = async (provider: 'google' | 'discord') => {
 		await authClient.signIn.social({ provider, callbackURL: '/dashboard' });
 	};
@@ -32,17 +27,7 @@ export const LoginButton = () => {
 				<Button>Login</Button>
 			</DropdownMenuTrigger>
 			<DropdownMenuContent className="min-w-fit *:cursor-pointer">
-				<Dialog open={open} onOpenChange={setOpen}>
-					<DialogTrigger asChild>
-						<DropdownMenuItem onSelect={(event) => event.preventDefault()}>
-							<EmailIcon />
-							<div>Email</div>
-						</DropdownMenuItem>
-					</DialogTrigger>
-					<DialogContent>
-						<EmailLoginForm onSuccess={() => setOpen(false)} />
-					</DialogContent>
-				</Dialog>
+				<EmailLogin />
 				<DropdownMenuItem onClick={() => login('google')} asChild>
 					<button>
 						<GoogleIcon />
